# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Library API",
        "contact": {
            "email": "bmpcorp9@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/album": {
            "get": {
                "tags": [
                    "Albums",
                    "List"
                ],
                "summary": "Get a paginated list of albums",
                "operationId": "9a80fc02ab790c1c8251b09a5c822993",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query for album/artist names or music genres",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "no_explicit",
                        "in": "query",
                        "description": "Filter out albums with explicit content (1 for true, 0 for false)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of albums retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Album Title"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "First album after reunition"
                                                            },
                                                            "genres": {
                                                                "type": "string",
                                                                "example": "Pop, Rock"
                                                            },
                                                            "year": {
                                                                "type": "integer",
                                                                "example": 2023
                                                            },
                                                            "song_count": {
                                                                "type": "integer",
                                                                "example": 9
                                                            },
                                                            "has_explicit_lyrics": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "cover": {
                                                                "type": "string",
                                                                "example": "https://example.com/album_cover.png"
                                                            },
                                                            "artist": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "title": {
                                                                        "type": "string",
                                                                        "example": "Artist Title"
                                                                    },
                                                                    "logo": {
                                                                        "type": "string",
                                                                        "example": "https://example.com/logo.png"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 100
                                                        },
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "pages": {
                                                            "type": "integer",
                                                            "example": 10
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/album/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get a single album by ID",
                "operationId": "e4d9a45c41321bb75b12788cf324492d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the album to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "item": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "title": {
                                                            "type": "string",
                                                            "example": "Album Title"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "First album after reunition"
                                                        },
                                                        "genres": {
                                                            "type": "string",
                                                            "example": "Pop, Rock"
                                                        },
                                                        "year": {
                                                            "type": "integer",
                                                            "example": 2023
                                                        },
                                                        "song_count": {
                                                            "type": "integer",
                                                            "example": 9
                                                        },
                                                        "has_explicit_lyrics": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "cover": {
                                                            "type": "string",
                                                            "example": "https://example.com/album_cover.png"
                                                        },
                                                        "artist": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "title": {
                                                                    "type": "string",
                                                                    "example": "Artist Title"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "This Band Rocks!"
                                                                },
                                                                "genres": {
                                                                    "type": "string",
                                                                    "example": "Pop, Rock"
                                                                },
                                                                "country_id": {
                                                                    "type": "integer",
                                                                    "example": 15
                                                                },
                                                                "logo": {
                                                                    "type": "string",
                                                                    "example": "https://example.com/logo.png"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}